pyproject.toml
[project]
name = "ai-config"
version = "0.1.0"
description = "Declarative plugin manager for Claude Code"
authors = [
    { name = "Alex Furrier", email = "afurrier@gmail.com" }
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "questionary>=2.0",
    "requests>=2.28",
    "rich>=13.0",
    "watchdog>=3.0",
]

[project.scripts]
ai-config = "ai_config.cli:main"

[project.optional-dependencies]
dev = [
    "ty>=0.0.2",
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0",
    "types-requests>=2.28",
]
docs = [
    "mkdocs-material>=9.6.14",
    "mkdocstrings[python]>=0.26.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_config"]

[tool.ruff]
line-length = 100
target-version = "py39"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]

ignore = [
    "E501",   # line too long (handled by line-length)
]

[tool.ruff.lint.isort]
known-first-party = ["ai_config"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.environment]
python-version = "3.9"

[tool.ty.src]
include = ["src"]

[tool.ty.terminal]
error-on-warning = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests",
]
